+++ src/pages/LikeCoin/index.tsx
@@ -1,0 +1,14 @@
+import * as React from 'react'
+
+import GenerateTempLikerIds from './GenerateTempLikerIds'
+import TransferLIKE from './TransferLIKE'
+
+const LikeCoin = () => {
+  return (
+    <>
+      <GenerateTempLikerIds />
+      <TransferLIKE />
+    </>
+  )
+}
+export default LikeCoin