+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2019-10-03
+
+### Added
+
+- Add OAuthClient list page
+- Add form to OAuthClient detail page
+- Add generate likerIds trigger
+- Add Add `<TransferLIKE>`
+
+### Changed
+
+- Sync schema
+
 ## [0.4.2] - 2019-09-03
 
 ### Added